+++ lib/src/widgets/center_pop_container.dart
@@ -1,0 +1,139 @@
+import 'package:flutter/material.dart';
+import 'package:hexcolor/hexcolor.dart';
+
+void Function() popCenterContainer({
+  required BuildContext context,
+  double extentHeight = 0,
+  String cancelText = 'Cancel',
+  String confirmText = 'Ok',
+  required List<Widget> children,
+  void Function()? onCancel,
+  bool Function()? onConfirm,
+}) {
+  const radius = Radius.circular(10);
+  final Color borderColor = Colors.grey[500]!;
+  final double extentNum = children.length - 2 > 0 ? children.length - 2 : 0;
+  final double height = extentNum * 33.33 + extentHeight;
+  final double padding = 10;
+
+  showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      final content = AlertDialog(
+        backgroundColor: Colors.white.withOpacity(0),
+        contentPadding: EdgeInsets.all(0),
+        content: Container(
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(radius),
+            color: HexColor('#F0F0F0'),
+          ),
+          height: 140 + height,
+          child: Column(
+            children: [
+              Container(
+                padding: EdgeInsets.all(padding),
+                width: double.infinity,
+                height: 90 + height,
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.spaceAround,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    for (int index = 0; index < children.length; index++)
+                      Wrap(
+                        children: [
+                          children[index],
+                          if (index + 1 != children.length)
+                            const Padding(padding: EdgeInsets.only(top: 10)),
+                        ],
+                      )
+                  ],
+                ),
+              ),
+              Container(
+                height: 50,
+                width: double.infinity,
+                decoration: BoxDecoration(
+                  border: Border(top: BorderSide(width: 1, color: borderColor)),
+                ),
+                child: LayoutBuilder(
+                  builder: (context, constraints) {
+                    final width = constraints.maxWidth / 2;
+                    return Row(
+                      children: [
+                        GestureDetector(
+                          behavior: HitTestBehavior.opaque,
+                          onTap: () {
+                            Navigator.pop(context);
+                            if (onCancel != null) onCancel();
+                            return;
+                          },
+                          child: SizedBox(
+                            width: width,
+                            height: constraints.maxHeight,
+                            child: Center(
+                              child: Text(
+                                cancelText,
+                                style: TextStyle(color: Colors.red),
+                              ),
+                            ),
+                          ),
+                        ),
+                        GestureDetector(
+                          behavior: HitTestBehavior.opaque,
+                          onTap: () {
+                            if (onConfirm != null) {
+                              final result = onConfirm();
+
+                              if (result) Navigator.pop(context);
+                            }
+                            return;
+                          },
+                          child: Container(
+                            width: width,
+                            decoration: BoxDecoration(
+                              border: Border(
+                                left: BorderSide(width: 1, color: borderColor),
+                              ),
+                            ),
+                            height: constraints.maxHeight,
+                            child: Center(
+                              child: Text(
+                                confirmText,
+                                style: TextStyle(color: HexColor('#3B6AF7')),
+                              ),
+                            ),
+                          ),
+                        )
+                      ],
+                    );
+                  },
+                ),
+              ),
+            ],
+          ),
+        ),
+
+        // actions: [
+        //   SizedBox(child: Text('Cancel')),
+        //   SizedBox(child: Text('Cancel')),
+        // ],
+        alignment: Alignment.center,
+      );
+
+      // Material(
+      //   child: Center(
+      //     child: Container(
+      //       width: MediaQuery.of(context).size.width - margin * 2,
+      //       height: 150,
+      //     ),
+      //   ),
+      // );
+
+      return content;
+    },
+  );
+  // AlertDialog(
+  //
+  // )
+  return () => Navigator.pop(context);
+}