+++ lib/wuchuheng_ui.dart
@@ -1,4 +1,5 @@
 library wuchuheng_ui;
 
 export 'src/on_bottom_sheet/index.dart';
+export 'src/on_confirm_dialog/index.dart';
 export 'src/on_dialog/index.dart';