+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.1
+
+- feat: create onBottomSheet function.
+- feat: create onDialog function.
+- feat: create onConfirmDialog function.